[Model] Support training of Qwen3-VL (#951)
### PR Category
<!-- One of [ Train | Inference | Compress | Serve | RL | Core |
Hardware | CICD | Tools | Others ] -->
Train
### PR Types
<!-- One of [ User Experience | New Features | Bug Fixes | Improvements
| Performance | Breaking Change| Deprecations | Test Case | Docs |
Others ] -->
New Features
### PR Description
<!-- Describe what you’ve done -->
This pull request mainly focuses on continuing to train the Qwen3-VL
model. The following are the key points covered in this PR:

- [ Dense Model] Support for distributed training, checkpoint
conversion, and simple data preprocessing of Qwen3-VL Model with the
Dense Architecture.

- [MoE Model] coming later

An accuracy comparison experiment was conducted between Llama-factory
and FlagScale. Under the same data and strategy, the initial loss
difference was in the thousandths(<0.01), and the convergence trend was
consistent.

#### How to use
You can refer to
[QuickStart](https://github.com/flagos-ai/FlagScale/pull/951/files#diff-d1f0f53dd1f17b6cff2ffaba83bccbfbeace211c977347e542b751029a9f8391)

#### Reference
This implementation in this PR is mainly referred from our Qwen2.5-VL
implementation of https://github.com/flagos-ai/FlagScale/pull/483

---------

Co-authored-by: lzy <lizhiyu211@foxmail.com>